+++ src/components/navigation/Navbar.tsx
@@ -17,7 +17,7 @@ import React, { useState } from "react";
 import {
   ImageColourSchemeSwitch,
   ImageColourSchemeSwitchType,
-} from "../controls/ImageColorSchemeSwitch";
+} from "../controls/ImageColourSchemeSwitch";
 import { Logo } from "../controls/Logo";
 
 interface NavLinksProps {