+++ src/index.ts
@@ -6,7 +6,7 @@ export * from "./components/navigation/Navbar";
 
 // components/controls
 export * from "./components/controls/ColourSchemeButton";
-export * from "./components/controls/ImageColorSchemeSwitch";
+export * from "./components/controls/ImageColourSchemeSwitch";
 export * from "./components/controls/Logo";
 export * from "./components/controls/User";
 export * from "./components/controls/VisitInput";